refactor(pivotal): split pivotal.py into load + apply modules
pivotal.py was 396 LOC, over the new quality_gates 300 LOC threshold. Split
into two sibling modules with a re-export shim to keep existing imports
working (engine.py, tests/test_pivotal.py reference imagecli.pivotal).

- pivotal_load.py (160 LOC): PivotalEmbedding, detect_pivotal_in_lora,
  _validate, load_pivotal_embedding, sanity caps
- pivotal_apply.py (237 LOC): apply_pivotal_to_pipe,
  _maybe_convert_prompt, _patch_encode_prompt
- pivotal.py (42 LOC): re-export shim preserving the public surface

All 29 pivotal tests pass. Behavior unchanged.

Refs #53
